eatgo/eatgo-customer-api/src/main/java/kr/co/fastcampus/eatgo/interfaces/UserService.java
package kr.co.fastcampus.eatgo.interfaces;

import kr.co.fastcampus.eatgo.domain.User;
import kr.co.fastcampus.eatgo.domain.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@Transactional
public class UserService {
    UserRepository userRepository;

    //의존성 설정
    @Autowired
    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User registerUser(String email, String name, String password) {
        User user = User.builder()
                .id(1004L)
                .email(email)
                .name(name)
                .password(password) //암호화가 필요
                .build();

        return userRepository.save(user);


    }
}
